修复LINGYUN.Abp.Tencent模块语法、单词拼写问题
