+++ aspnet-core/modules/cloud-tencent/LINGYUN.Abp.Sms.Tencent/LINGYUN/Abp/Sms/Tencent/Settings/TencentCloudSmsSettingProvider.cs
@@ -8,10 +8,10 @@ namespace LINGYUN.Abp.Sms.Tencent.Settings
     {
         public override void Define(ISettingDefinitionContext context)
         {
-            context.Add(CreateAliyunSettings());
+            context.Add(CreateTencentSettings());
         }
 
-        private SettingDefinition[] CreateAliyunSettings()
+        private SettingDefinition[] CreateTencentSettings()
         {
             return new SettingDefinition[]
             {
