+++ aspnet-core/modules/cloud-tencent/LINGYUN.Abp.Sms.Tencent/Volo/Abp/Sms/TencentSmsSenderExtensions.cs
@@ -14,7 +14,7 @@ namespace Volo.Abp.Sms
         /// <param name="phoneNumber">发送手机号</param>
         /// <param name="templateParams">短信模板参数</param>
         /// <returns></returns>
-        public static async Task SendAsync(
+        public async static Task SendAsync(
             this ISmsSender smsSender,
             string templateCode,
             string phoneNumber,
@@ -38,7 +38,7 @@ namespace Volo.Abp.Sms
         /// <param name="phoneNumber">发送手机号</param>
         /// <param name="templateParams">短信模板参数</param>
         /// <returns></returns>
-        public static async Task SendAsync(
+        public async static Task SendAsync(
             this ISmsSender smsSender,
             string signName,
             string templateCode,
