--- aspnet-core/modules/cloud-tencent/LINGYUN.Abp.Tencent.SettingManagement/LINGYUN/Abp/Tencent/SettingManagement/ITenantCloudSettingAppService.cs
@@ -1,7 +1,0 @@
-﻿using LINGYUN.Abp.SettingManagement;
-
-namespace LINGYUN.Abp.Tencent.SettingManagement;
-
-public interface ITenantCloudSettingAppService : IReadonlySettingAppService
-{
-}
