+++ aspnet-core/modules/cloud-tencent/LINGYUN.Abp.Tencent.SettingManagement/LINGYUN/Abp/Tencent/SettingManagement/ITencentCloudSettingAppService.cs
@@ -1,0 +1,7 @@
+﻿using LINGYUN.Abp.SettingManagement;
+
+namespace LINGYUN.Abp.Tencent.SettingManagement;
+
+public interface ITencentCloudSettingAppService : IReadonlySettingAppService
+{
+}
